util/packet: add helpers for UDP emulation on top of AF_PACKET sockets
This allows us to send/recevie UDP packets without the IP stack being
fully configured. It is not possible to neither send nor receive a
UDP packet using an AF_INET socket when the destination IP has not
been configured. We therefore emulate both sending and receiving of
UDP packets on the AF_PACKET layer, as well as binding an AF_PACKET
socket to what ammounts to a DHCP transaction (rather than to an
address as one would do if the IP stack had been configured).

Signed-off-by: Tom Gundersen <teg@jklm.no>