message: add message handling
This hides the fact that DHCP options are not written in a linear buffer,
and exposes them as if they were. For outgoing messages, options can
be appended, and for incoming messages options can be queried. In both
cases both the DHCPv4 header and the whole raw message are exposed in
the API.

Signed-off-by: Tom Gundersen <teg@jklm.no>