test: decouple netns creation from veth creation
Introduce a funciton, test_netns_new(), which returns a new network
namespace as a file descriptor. Allow this to be passed into
test_veth_new(), rather than the veth constructor implicitly creating
a pair of new namespaces too.

This allows veth links to be freely created in any namespaces, not
just new ones.

Signed-off-by: Tom Gundersen <teg@jklm.no>