network: filter all sockets the same
Only return BOOTREPLY/BOOTREQUEST messages (as appropriate), with
a full DHCP header and a valid magic cookie.

The receiver still must verify that the packet is destined for them
(using transaction id, client id or server id, as appropriate).

Before this patch, the packet sockets had a stricter filter than
the UDP sockets, this allows the userspace verification to be the
same for all the socket types.

Signed-off-by: Tom Gundersen <teg@jklm.no>