network/server/udp: drop SO_REUSEADDR
We want to be able to run one DHCP server instance per interface, from the
documentation it appears that we require SO_REUSEADDR in order to bind to
INADDR_ANY and the same port several times on the same host. However, it
turns out that as long as each instance is bound to a different interface
with SO_BINDTODEVICE, this is not necessary.

Signed-off-by: Tom Gundersen <teg@jklm.no>