connection: force-set MTU option
Make sure we always set the MTU option. There are server
implementations that interpret missing MTU-options to mean any packet
size is supported. This is definitely not what we want as default
behavior, so lets pick the minimum required by the IP spec.

Note that in case we want the mentioned behavior, we can always add
options to suppress the MTU field. However, we don't want to tie the
fact that a caller might not have set our MTU to omitting the MTU
option. This should be explicit, if we want it.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>