socket/server: set the source IP in outgoing UDP packets
The server supports having several IP addresses configured, but it identifies itself
with one of them as the SERVER_IDENTIFIER option. For the sake of consistency, make
it possible to specify the source address when sending UDP packets from the server,
so this can be set to the address contained in the payload.

Signed-off-by: Tom Gundersen <teg@jklm.no>