connection: set secs based on start and send time
Maintain the secs value internally. We set it to be one more than
the number of seconds between start and send. The reason to add
one is to avoid ever setting the value to zero, as some servers
do not support that.

A SELECT message always reuses the secs value from the corresponding
DISCOVER, and never increments it after that.

Signed-off-by: Tom Gundersen <teg@jklm.no>