client-connection: don't reset the xid on SELECT resend
We must always use the xid from OFFER in SELECT, as we want the
header to be the same (in case proxies/servers change their
behavior based on the header, even though that seems very unlikely).

Moreover, we want to accept any ACK in response to a SELECT, even if
it arrives as a late response to a previous message. The reason is
that servers are allowed to take their time verifying the lease before
acking it.

As a result of this, we never know which SELECT an ACK is in response to
(when it is in response to an ACK), so in that case we do not attach the
send_time to the ACK, as we do not know which one to use.

Signed-off-by: Tom Gundersen <teg@jklm.no>