server: introduce tracking objects for IP addresses
Rather than referencing the addresses directly, return a tracking
object when an address is added, and automatically drop the
address from the server when the object is freed.

At the moment we still only support one address, but the API is
now made to support any number of addresses once we add an
RBTree to manage them.

Signed-off-by: Tom Gundersen <teg@jklm.no>