client/probe: query current time early in dispatcher
Rather than querying the clock all over the place, we query it early
and pass it through by value.

We still have to make sure to query it *AFTER* calling into read(2), as
otherwise we could possible miss a wakeup event for edge-triggered
timers (which we don't have, but lets prepare..).

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>