test/run: provide DHCP client test run
This adds a new test ./test-run-client, which runs a full DHCP client
and accepts all kinds of tweaks on the command-line. It is mainly meant
for testing during development, but also accepts the --test
command-line to be run as part of the test-suite.

This should never be exported! It provides internal tweaks and is not
meant for production.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>