socket: expose the destination address for incoming server packets
This allows us to distinguish between broadcast and unicast on the
server (which is needed if we want to know if a given packet
is RENEW or REBIND). This might not be necessary, but the cost is
very minor, and it would make for a more complete API with better
test coverage.

Signed-off-by: Tom Gundersen <teg@jklm.no>