server-connection: remember the high-level packet type in incoming packets
The REQUEST packet type is overloaded, so this deduces the high-level
packt type for all incomnig packets, so we can use that directly on
the server side, just like we do on the client side.

We distinguish between all the packet types to perfectly match the
client API, even though some types strictly speaking do not need to
be distinguished.

Signed-off-by: Tom Gundersen <teg@jklm.no>