client: integrate connections with probes
This deduplicates the configuration data in connections, in favor of
taking a pointer to a ClientConfig object. This way, the connection
state gets a lot simpler.

Furthermore, this integrates the epoll-fd with probes. We now take it
in the constructor, and require it to be valid the entire lifetime of
the connection. However, we also allow -1 to be passed, making the
entire connection a detached probe, which allows us to keep the error
handling async.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>