probe: add GRANTED state
When a lease is granted, we no longer immediately enter the BOUND state,
but rather stay in the new GRANTED state until the caller calls _accept().

In this new state we should not be sending out any packets, so if T1 or T2
triggers they will only be processed after the caller has accepted the
lease. However, if the lease expires, we still forward that event to the
caller.

Signed-off-by: Tom Gundersen <teg@jklm.no>