client/connection: avoid state-dependence in message creation
When creating messages, we cannot predict at what time they are sent.
Therefore, we must not make them dependent on connection state.
Instead, every message must only depend on its own parameters.

Make sure we do not switch on the current socket-state, but instead use
our knowledge about the message type to predict on which socket it must
be sent.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>