client: arm timers after starting a probe
Usually we re-arm the timers after dispatch, but in order to catch
the timers that might have been armed when the probe was created,
we must also re-arm them in client_probe(). Note that we might want
to cancel timeouts in probe_free(), but we don't. Worst case we have
one spurious timer that notices there is nothing to do and cancels
further timeouts.

Signed-off-by: Tom Gundersen <teg@jklm.no>