lease: fix linking between leases and probes
A lease may or may not be linked into a probe. The lease does not
pin the probe, nor does the probe pin the lease. It is the
responsibility of the probe to unink all leases when it is freed,
and similarly the responsibility of a lease to unlink itself when
it is freed.

Signed-off-by: Tom Gundersen <teg@jklm.no>