client/connection: only request broadcast reply when IP is not configured
When the IP stack has not been configured, some hardware is unable
to receive unicast IP packets. When that is the case, the caller
must explicitly request broadcasts from the server.

However, we never want to pass the BROADCAST bit when the IP address
has been configured, as that is never necessary. Change the logic
accordingly.

Signed-off-by: Tom Gundersen <teg@jklm.no>