test/run/client: fall back to explicitly request the options we use
We do not support (but do not yet enforce) querying lease options that
were not explicitly configured. If the caller provides the list it is
their responsibilty to include what is needed (since this is a debug
tool we allow them to decide the order etc), but otherwise we just
provide the ones that are required.

Signed-off-by: Tom Gundersen <teg@jklm.no>