n-dhcp4/socket: use SO_REUSEADDR on UDP socket
Otherwise, other applications cannot bind to port 0.0.0.0:68 at the same time.
This is for example what dhclient wants to do. So even when running
dhclient on another, unrelated interface, it would fail to bind the UDP
socket and quit.

Note that also systemd-networkd's DHCPv4 client sets this socket option.
Presumably for the same reasons.

https://gitlab.freedesktop.org/NetworkManager/NetworkManager/-/commit/cf7662bc52da849eb492d0fa710d553e3bbafa58

https://github.com/nettools/n-dhcp4/pull/12

Signed-off-by: Thomas Haller <thaller@redhat.com>