n-dhcp4: support init-reboot state
Currently the client always starts from the INIT state (i.e. sending a
discover message). If a requested-ip was specified by the caller, it
is added as an option in the discover.

The RFC allows to skip the discover by entering the INIT-REBOOT state
and starting directly with a broadcast request message containing the
requested IP address. Implement that.

https://gitlab.freedesktop.org/NetworkManager/NetworkManager/-/commit/af03b779804f8c0b7efff548691921ae8c29c77a
https://gitlab.freedesktop.org/NetworkManager/NetworkManager/-/commit/218782a9a3c326f5c8cc3ea40dc0cd039060b188
https://gitlab.freedesktop.org/NetworkManager/NetworkManager/-/commit/3a9b069c4161dbfc652e1b6b5447790165da6d87
https://gitlab.freedesktop.org/NetworkManager/NetworkManager/-/commit/de818bf610ba0fed2b275a19fe3e6f9cbfa1862b
https://gitlab.freedesktop.org/NetworkManager/NetworkManager/-/commit/4bcdc3c1ebe3e2e8a967ff067ecb2a8cbfc2f6ab

https://github.com/nettools/n-dhcp4/pull/14

Signed-off-by: Beniamino Galvani <bgalvani@redhat.com>
Signed-off-by: Thomas Haller <thaller@redhat.com>