n-dhcp4: use packet socket in rebinding state
After t1, the client tries to renew the lease by contacting via the
udp socket the server specified in the server-id option. If this
fails, after t2 it tries to contact any server using broadcast. For
this to work, the packet socket must be used or reply packets from a
server different from the one specified in the server-id option will
be discarded.

https://gitlab.freedesktop.org/NetworkManager/NetworkManager/-/commit/f860e929c04ef27ffb16eb7fb5456123a8595d2c

https://github.com/nettools/n-dhcp4/pull/15

Signed-off-by: Beniamino Galvani <bgalvani@redhat.com>
Signed-off-by: Thomas Haller <thaller@redhat.com>