probe: move back to INIT after lease expires
Move back to INIT state after the lease expires, as per section 4.4.5
of RFC 2131. Previously the client just moved to EXPIRED, closed the
connection and cleared the probe, leaving to the caller of the library
the choice to create a new client instance and to start from
scratch. However, it seems more useful that the client, once
initialized, always tries to get a lease even after an expiration.

https://gitlab.freedesktop.org/NetworkManager/NetworkManager/-/commit/182a8021f39c943af87d2bfc703d09dbc7e25bf7

Signed-off-by: Thomas Haller <thaller@redhat.com>