packed: use c_errno() in packet_sendto_udp() to avoid compiler warning
With LTO enabled, the compiler might think that "len" in n_dhcp4_socket_packet_send()
might be uninitialized. That is even a correct assumption, as the compiler does not
understand the API of sendmsg() and that sendmsg() is supposed to set a positive errno.

Work around by using c_errno() in packet_sendto_udp().

    src/n-dhcp4-c-connection.c: In function n_dhcp4_c_connection_send_request:
    src/n-dhcp4-socket.c:368:19: error: len may be used uninitialized in this function [-Werror=maybe-uninitialized]
             } else if (len != n_buf) {
                       ^
    src/n-dhcp4-socket.c:351:23: note: len was declared here
             size_t n_buf, len;
                           ^

https://github.com/nettools/n-dhcp4/pull/22