Add a check-dhcp4 optional binary
This tool is a standalone binary that can use to check
if a DHCP server is "alive", it simply broadcast a DHCP DISCOVER
request and listen for an OFFER or timed out.
It doesn't accept the lease offered by the server.

This can be useful to check if a DHCP server is responsive before
waiting for the future lease renew, it can be part
of a monitoring tool which watch some IT infrastructure.
For example NAGIOS has something similar as a plugin
(see https://nagios-plugins.org/doc/man/check_dhcp.html)

Signed-off-by: Frederic Martinsons <frederic.martinsons@unabiz.com>